% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_max.R
\name{enough_items}
\alias{enough_items}
\title{Function to check there are at least certain number of items per factor}
\usage{
enough_items(loadings, min_items = 3, eps = 1e-08)
}
\arguments{
\item{loadings}{Loading matrix resulting from SEAFA.}

\item{min_items}{Minimum number of items per factors.}

\item{eps}{Numerical tolerance used to treat very small loadings as zero.}
}
\value{
A logical vector indicating whether each factor has at least \code{min_items} nonzero loadings.
}
\description{
Function to check there are at least certain number of items per factor
}
\examples{
\dontrun{
check <- enough_items(
  loadings)
}
}
