% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_retention.R
\name{factor_small}
\alias{factor_small}
\title{Function to retain number of factors/components for traditional size data set using various methods.}
\usage{
factor_small(data)
}
\arguments{
\item{data}{A NxJ matrix of observed variables.}
}
\value{
Number of factors/components.
}
\description{
Function to retain number of factors/components for traditional size data set using various methods.
}
\examples{
\dontrun{
data(big5, package = "qgraph")
X <- scale(big5)
a <- factor_small(X)
}
}
