% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_max.R
\name{fit_once}
\alias{fit_once}
\title{Function to fit SEAFA with LASSO penalty once and get loading matrix}
\usage{
fit_once(
  data,
  nfactors,
  lambda,
  maxiter = 50,
  eps = 10^-4,
  INIT = "semirational",
  nstarts = 1,
  standardize = TRUE
)
}
\arguments{
\item{data}{A data frame or matrix (NxJ). Data are always centered internally.}

\item{nfactors}{Number of factors.}

\item{lambda}{LASSO penalty parameter.}

\item{maxiter}{Maximum iterations for SEAFA with LASSO.}

\item{eps}{Convergence criterion based on difference in loss between iterations.}

\item{INIT}{Method to initialize loading matrix.}

\item{nstarts}{Number of starts.}

\item{standardize}{Logical. If \code{TRUE} (default), items are scaled to unit variance (N denominator)
after centering. Centering is applied regardless.}
}
\value{
Estimated loading matrix.
}
\description{
Function to fit SEAFA with LASSO penalty once and get loading matrix
}
\examples{
\dontrun{
P <- seafar_lasso_multistart(
  scale(big5),
  nfactors = 5,
  lambda = 10,
  INIT = 'mixed',
  nstarts = 1)
}
}
