% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{hardthr}
\alias{hardthr}
\title{Function for hard thresholding.}
\usage{
hardthr(loadings, C_c)
}
\arguments{
\item{loadings}{Loadings matrix JxR.}

\item{C_c}{Number of zero loadings.}
}
\value{
Loadings with exactly C_c nonzero loadings.
}
\description{
Function for hard thresholding.
}
\examples{
\dontrun{
hardthr(loadings, C_c)
}
}
\author{
Katrijn Van Deun
}
