% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.seafar.R
\name{is.seafar_original}
\alias{is.seafar_original}
\title{Model selection for SEAFAR using the Index of Sparseness.}
\usage{
is.seafar_original(
  data,
  nfactors,
  C,
  maxiter = 50,
  eps = 10^-4,
  INIT,
  orthogonal = FALSE,
  nstarts,
  standardize = TRUE
)
}
\arguments{
\item{data}{A data frame or matrix (NxJ).}

\item{nfactors}{Number of factors.}

\item{C}{Number of nonzero loadings.}

\item{maxiter}{Maximum number of iterations for the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{INIT}{Method to initialize loading matrix.}

\item{orthogonal}{Orthogonal or non-orthogonal factors, default is FALSE.}

\item{nstarts}{Number of starts.}

\item{standardize}{Logical. If \code{TRUE} (default), items are scaled to unit variance
after centering. Centering is applied regardless.}
}
\value{
A list with
\item{value}{IS value.}
\item{vaf}{Variance accounted for.}
\item{propzero}{Proportion of zero loadings.}
\item{smallestP}{Smallest nonzero loading.}
\item{maxsdP}{Maximum standard deviation of loading.}
\item{center}{Item means used to center the data.}
\item{scale}{Item standard deviations used to scale the data (1s if \code{standardize = FALSE}).}
}
\description{
Model selection for SEAFAR using the Index of Sparseness.
}
\examples{
\dontrun{
IS <- is.seafar_original(
  data = X,
  nfactors = 5,
  C = 240,
  INIT = "random",
  orthogonal = TRUE,
  nstarts = 50
)
}

}
