% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.seafar_lasso.R
\name{IS_lasso}
\alias{IS_lasso}
\title{Function to calculate the IS for each LASSO penalty parameter.}
\usage{
IS_lasso(
  data,
  nfactors,
  lambda,
  maxiter = 50,
  eps = 10^-4,
  INIT = "semirational",
  nstarts,
  standardize = TRUE
)
}
\arguments{
\item{data}{A data frame or matrix (NxJ). Data are always centered internally.}

\item{nfactors}{The number of factors.}

\item{lambda}{LASSO penalty tuning parameter.}

\item{maxiter}{Maximum number of iterations of the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{INIT}{Method to initialize loadings.}

\item{nstarts}{Number of starts.}

\item{standardize}{Logical. If \code{TRUE} (default), items are scaled to unit variance (N denominator)
after centering. Centering is applied regardless.}
}
\value{
A list with
\item{value}{IS value.}
\item{vaf}{Variance accounted for.}
\item{propzero}{Proportion of zero loadings.}
\item{smallestP}{Smallest nonzero loading.}
\item{maxsdP}{Maximum standard deviation of loading.}
}
\description{
Function to calculate the IS for each LASSO penalty parameter.
}
\examples{
\dontrun{
IS <- IS_lasso(
  data = X,
  nfactors = 5,
  lambda = 100,
  INIT = "mixed",
  nstarts = 50
)
}
}
