% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{LOSS}
\alias{LOSS}
\title{Title Loss function for LASSO}
\usage{
LOSS(data, scores, loadings, lambda)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{scores}{Factor scores matrix NxQ.}

\item{loadings}{Loadings matrix JxQ.}

\item{lambda}{LASSO penalty tuning parameter.}
}
\value{
Loss value when using LASSO penalty.
}
\description{
Title Loss function for LASSO
}
\examples{
\dontrun{
LOSS(X, scores, loadings, lambda)
}
}
