% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{oblprocr}
\alias{oblprocr}
\title{Function to calculate the solution to the olique procrustes problem.}
\usage{
oblprocr(data, scores, loadings, maxiter, eps)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{loadings}{Loadings matrix JxQ.}

\item{maxiter}{Maximum number of iterations.}

\item{eps}{Tolerance for difference in loss.}
}
\value{
Orthogonal factor scores matrix NxJ.
}
\description{
Function to calculate the solution to the olique procrustes problem.
}
\examples{
\dontrun{
orthprocr(X, loadings)
}
}
\author{
Katrijn Van Deun.
}
