% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{orthprocr}
\alias{orthprocr}
\title{Function to calculate the solution to the orthogonal procrustes problem.}
\usage{
orthprocr(data, loadings)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{loadings}{Loadings matrix JxQ.}
}
\value{
Orthogonal factor scores matrix NxJ.
}
\description{
Function to calculate the solution to the orthogonal procrustes problem.
}
\examples{
\dontrun{
orthprocr(X, loadings)
}
}
\author{
Katrijn Van Deun.
}
