% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pstr}
\alias{pstr}
\title{Partially specified target rotation.}
\usage{
pstr(loadings, target, weights, maxiter = 50, eps = 1e-04)
}
\arguments{
\item{loadings}{Matrix that needs to be rotated.}

\item{target}{Target towards which to rotated.}

\item{weights}{elementwise weights for rotation.}

\item{maxiter}{Maximum number of iterations of the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}
}
\value{
Rotation matrix and loss of the pstr criterion.
}
\description{
Partially specified target rotation.
}
\examples{
\dontrun{
ocean_pstr <- pstr(pcaloadings, target, weights)
}
}
