% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pstrLoss}
\alias{pstrLoss}
\title{Function to calculate the objective function associated with pstr.}
\usage{
pstrLoss(B, Tmat, target, weights)
}
\arguments{
\item{B}{Rotation matrix.}

\item{Tmat}{Original loading matrix.}

\item{target}{Target matrix for rotation.}

\item{weights}{Weight.}
}
\value{
Loss value.
}
\description{
Function to calculate the objective function associated with pstr.
}
\examples{
\dontrun{
loss <- pstrLoss(B, Tmat, target, weights)
}
}
