% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{reflexmat}
\alias{reflexmat}
\title{Function to construct matrix of reflections.}
\usage{
reflexmat(m)
}
\arguments{
\item{m}{Number of factors.}
}
\value{
Matrix of reflections.
}
\description{
Function to construct matrix of reflections.
}
\examples{
\dontrun{
matrix <- reflexmat(5)
}
}
\author{
Katrijn Van Deun and Zhengguo Gu.
}
