% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_autism_dataset.R
\name{reproduce_autism_dataset}
\alias{reproduce_autism_dataset}
\title{Reproduce the autism dataset from GEO accession GSE7329}
\usage{
reproduce_autism_dataset(
  source_url =
    "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE7nnn/GSE7329/matrix/GSE7329_series_matrix.txt.gz",
  envir = parent.frame()
)
}
\arguments{
\item{source_url}{Source URL of the original Series Matrix file on the NCBI GEO ftp server.}
}
\value{
Invisibly returns the processed matrix \code{autism}.
}
\description{
Downloads the original Series Matrix file from NCBI GEO, reproduces the
existing preprocessing steps.
}
\details{
Running \code{autism <- reproduce_autism_dataset()} is equivalent to \code{data("autism")}

For more information, consult \code{data/README.md}.
}
