% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{seafar_init_l1}
\alias{seafar_init_l1}
\title{Function for obtaining an initial loading matrix for LASSO penalty.}
\usage{
seafar_init_l1(data, nfactors, INIT)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{nfactors}{Number of factors.}

\item{INIT}{Methods to initialize loadings c("svd", "random", "mixed).}
}
\value{
A matrix of initial loadings.
}
\description{
Function for obtaining an initial loading matrix for LASSO penalty.
}
\examples{
\dontrun{
initialP_lasso <- seafar_init_l1(ocean, 5, INIT = "svd")
}
}
