% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ssres}
\alias{ssres}
\title{Loss function.}
\usage{
ssres(data, scores, loadings)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{scores}{Factor scores matrix NxQ.}

\item{loadings}{Loadings matrix JxQ.}
}
\value{
Sum of squared residuals.
}
\description{
Loss function.
}
\examples{
\dontrun{
ssres(X, scores, loadings)
}
}
\author{
Katrijn Van Deun.
}
