% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.seafar.R
\name{summary.is_seafar}
\alias{summary.is_seafar}
\title{Display a summary of the results of \code{is.seafar()}.}
\usage{
\method{summary}{is_seafar}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'is_seafar'.}

\item{...}{Argument to be passed to or from other methods.}
}
\value{
Summary of the results.
}
\description{
Display a summary of the results of \code{is.seafar()}.
}
\examples{
\dontrun{
summary(IS_result)
}
}
