% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.seafar_wst.R
\name{summary.ISwarmstart}
\alias{summary.ISwarmstart}
\title{Display a summary of the results of \code{is.seafar_wst()}.}
\usage{
\method{summary}{ISwarmstart}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'ISwarmstart'.}

\item{...}{Argument to be passed to or from other methods.}
}
\value{
Summary of the results.
}
\description{
Display a summary of the results of \code{is.seafar_wst()}.
}
\examples{
\dontrun{
summary(IS_result)
}
}
