% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nfactors.R
\name{summary.nfactors}
\alias{summary.nfactors}
\title{Display a summary of the results of \code{factor_retention()}.}
\usage{
\method{summary}{nfactors}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'factor_retention'.}

\item{...}{Argument to be passed to or from other methods.}
}
\value{
Summary of the number of factors retained using various methods.
}
\description{
Display a summary of the results of \code{factor_retention()}.
}
\examples{
\dontrun{
results <- factor_retention(X)
summary(results)
}
}
