% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_retention.R
\name{factor_retention}
\alias{factor_retention}
\title{Function to retain number of factors using various methods.}
\usage{
factor_retention(data)
}
\arguments{
\item{data}{A NxJ matrix of standardized items.}
}
\value{
Number of factors.
}
\description{
Function to retain number of factors using various methods.
}
\examples{
\dontrun{
data(big5, package = "qgraph")
X <- scale(big5)
a <- factor_retention(X)
}
}
