% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.seafar.R
\name{is.seafar}
\alias{is.seafar}
\title{Model selection for SEAFAR using modified Index of Sparseness
(Van Deun et al., 2025).}
\usage{
is.seafar(
  data,
  nfactors,
  maxiter = 50,
  eps = 10^-4,
  INIT,
  orthogonal = FALSE,
  nstarts,
  TOL,
  THR
)
}
\arguments{
\item{data}{A NxJ matrix of standardized items.}

\item{nfactors}{Number of factors.}

\item{maxiter}{Maximum number of iterations for the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{INIT}{Method to initialize loading matrix.}

\item{orthogonal}{Orthogonal or non-orthogonal factors, default is FALSE.}

\item{nstarts}{Number of starts.}

\item{TOL}{Decimal points for the loadings.}

\item{THR}{Thresholding selection for the smallest nonzero loading.}
}
\value{
The selected cardinality value.
}
\description{
Model selection for SEAFAR using modified Index of Sparseness
(Van Deun et al., 2025).
}
\examples{
\dontrun{
selcard <- is.seafar(
  data = X,
  nfactors = 5,
  INIT = "random",
  orthogonal = TRUE,
  nstarts = 50,
  TOL = 2,
  THR = 0.15
)
}
}
