% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.seafar_lasso.R
\name{is.seafar_lasso}
\alias{is.seafar_lasso}
\title{Model selection procedure with IS to choose the LASSO penalty paramter.}
\usage{
is.seafar_lasso(
  data,
  nfactors,
  max_lambda,
  maxiter = 50,
  eps = 10^-4,
  INIT,
  nstarts
)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{nfactors}{The number of factors.}

\item{max_lambda}{Maximum LASSO penalty parameter.}

\item{maxiter}{Maximum number of iterations of the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{INIT}{Method to initialize loading matrix.}

\item{nstarts}{Number of starts.}
}
\value{
The selected LASSO penalty parameter.
}
\description{
Model selection procedure with IS to choose the LASSO penalty paramter.
}
\examples{
\dontrun{
lambda <- is.seafar_lasso(
  data = X,
  nfactors = 5,
  max_lambda = 100,
  INIT = 'mixed',
  nstarts = 50)
}
}
