% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.seafar_lasso.R
\name{is.seafar_lasso}
\alias{is.seafar_lasso}
\title{Model selection for SEAFAR with LASSO penalty using the Index of Sparseness.}
\usage{
is.seafar_lasso(data, nfactors, lambda, maxiter, eps, INIT, nstarts)
}
\arguments{
\item{data}{A data frame containing the dataset (standardized).}

\item{nfactors}{The number of factors.}

\item{lambda}{LASSO penalty tuning parameter.}

\item{maxiter}{Maximum number of iterations of the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{INIT}{Method to initialize loadings.}
}
\value{
A list with
\item{value}{IS value.}
\item{vaf}{Variance accounted for.}
\item{propzero}{Proportion of zero loadings.}
\item{smallestP}{Smallest nonzero loading.}
\item{maxsdP}{Maximum standard deviation of loading.}
}
\description{
Model selection for SEAFAR with LASSO penalty using the Index of Sparseness.
}
\examples{
\dontrun{
IS <- is.seafar_lasso(
  data = X,
  nfactors = 5,
  lambda = 100,
  INIT = "mixed",
  nstarts = 50
)
}

}
