% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.seafar_wst.R
\name{is.seafar_wst}
\alias{is.seafar_wst}
\title{Model selection for SEAFAR using Index of Sparseness with warm start}
\usage{
is.seafar_wst(
  data,
  nfactors,
  eps = 1e-04,
  maxiter = 50,
  INIT,
  orthogonal = FALSE,
  card.length,
  show_progress = FALSE
)
}
\arguments{
\item{data}{A NxJ matrix of standardized items.}

\item{nfactors}{Number of factors.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{maxiter}{Maximum number of iterations for the AO procedure.}

\item{INIT}{Method to initialize loading matrix.}

\item{orthogonal}{Orthogonal or non-orthogonal factors, default is FALSE.}

\item{card.length}{Number of candidate cardinality values to test.}

\item{show_progress}{Option of print progress bar, default is FALSE.}
}
\value{
A list with
\item{value}{IS value.}
\item{vaf}{Variance accounted for.}
\item{propzero}{Proportion of zero loadings.}
\item{smallestP}{Smallest nonzero loading.}
}
\description{
Model selection for SEAFAR using Index of Sparseness with warm start
}
\examples{
\dontrun{
IS_result <- is.seafar_wst(
  data = X,
  nfactors = 5,
  INIT = 'random',
  orthogonal = TRUE,
  card.length = 100)
}
}
