% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_max.R
\name{lambda_max}
\alias{lambda_max}
\title{Function to find the maximum LASSO penalty parameter for a given data set}
\usage{
lambda_max(
  data,
  nfactors,
  maxiter = 50,
  eps = 10^-4,
  nstarts = 1,
  min_items = 3,
  TOL = 10^-4,
  INIT,
  max_iter = 25
)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{nfactors}{Number of factors.}

\item{maxiter}{Maximum iterations for SEAFA with LASSO.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{nstarts}{Number of starts}

\item{min_items}{Number of minimum items per factor.}

\item{TOL}{Tolerance for the binary search convergence criterion.}

\item{INIT}{Method to initialize loading matrix.}

\item{max_iter}{Maximum iterations for the binary search.}
}
\value{
The maximum value for LASSO penalty parameter.
}
\description{
Function to find the maximum LASSO penalty parameter for a given data set
}
\examples{
\dontrun{
max_lambda <- lambda_max(
  data = scale(big5),
  nfactors = 5,
  INIT = 'mixed')
}
}
