% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar.R
\name{seafar}
\alias{seafar}
\title{Sparse exploratory approximate factor analysis}
\usage{
seafar(
  data,
  nfactors,
  C,
  eps = 1e-04,
  maxiter = 50,
  initloadings = NULL,
  INIT,
  orthogonal = FALSE,
  standardize = TRUE
)
}
\arguments{
\item{data}{A data frame or matrix (NxJ). Data are always centered internally.}

\item{nfactors}{The number of factors}

\item{C}{The number of nonzero loadings}

\item{eps}{Convergence criterion based on difference in loss between iterates}

\item{maxiter}{Maximum number of iterations of the AO procedure}

\item{initloadings}{Initial loading matrix if available.}

\item{INIT}{Method to initialize loadings}

\item{orthogonal}{Orthogonal or non-orthogonal factors, default is FALSE.}

\item{standardize}{Logical. If \code{TRUE} (default), items are scaled to unit variance (N denominator)
after centering. Centering is applied regardless.}
}
\value{
Factor loading and factor score matrices, plus the \code{center} and \code{scale} used to transform the data.
}
\description{
Function for exploratory approximate factor analysis resulting in a sparse
measurement model
}
\examples{
\dontrun{
big5_seafa <- seafar(USArrests, 2, 4, INIT = "svd", orthogonal = TRUE)
}
}
