% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar.R
\name{seafar}
\alias{seafar}
\title{Sparse exploratory approximate factor analysis}
\usage{
seafar(
  data,
  nfactors,
  C,
  eps = 1e-04,
  maxiter = 50,
  initloadings = NULL,
  INIT,
  orthogonal = FALSE
)
}
\arguments{
\item{data}{A data frame containing the dataset (standardized).}

\item{nfactors}{The number of factors}

\item{C}{The number of nonzero loadings}

\item{eps}{Convergence criterion based on difference in loss between iterates}

\item{maxiter}{Maximum number of iterations of the AO procedure}

\item{initloadings}{Initial loading matrix if available.}

\item{INIT}{Method to initialize loadings}

\item{orthogonal}{Orthogonal or non-orthogonal factors, default is FALSE.}
}
\value{
Factor loading and factor score matrices
}
\description{
Function for exploratory approximate factor analysis resulting in a sparse
measurement model
}
\examples{
\dontrun{
big5_seafa <- seafar(ocean, 5, 240, INIT = "svd", orthogonal = TRUE)
}
}
