% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar.R
\name{seafar_general}
\alias{seafar_general}
\title{Function for exploratory approximate factor analysis resulting in a sparse
measurement model for non-orthogonal factors.}
\usage{
seafar_general(
  data,
  nfactors,
  C,
  eps = 1e-04,
  maxiter = 50,
  initloadings,
  INIT
)
}
\arguments{
\item{data}{A NxJ matrix of standardized items.}

\item{nfactors}{Number of factors.}

\item{C}{Number of nonzero loadings.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{maxiter}{Maximum number of iterations of the AO procedure.}

\item{initloadings}{Initial loading matrix if available.}

\item{INIT}{Method to initialize loadings.}
}
\value{
\item{loadings}{The best estimated loading matrix.}
\item{scores}{The best estimated factor score matrix.}
\item{PVE}{A vector of PVE in each iteration of the AO procedure.}
}
\description{
Function for exploratory approximate factor analysis resulting in a sparse
measurement model for non-orthogonal factors.
}
\examples{
\dontrun{
big5_seafa <- seafar(ocean, 5, 240, INIT = "svd")
}
}
