% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{seafar_init}
\alias{seafar_init}
\title{Function for obtaining an initial loading matrix.}
\usage{
seafar_init(data, nfactors, C, INIT)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{nfactors}{The number of factors Q.}

\item{C}{The number of zero loadings.}

\item{INIT}{Methods to initialize loadings c("svd", "random", "rational", "mixed).
Rational start uses the result of Ding and He (2004) to partition the
variables into nfactors clusters.
Mixed start combines random (20\%) and SVD (80\%).}
}
\value{
A matrix of initial loadings.
}
\description{
Function for obtaining an initial loading matrix.
}
\examples{
\dontrun{
initialP <- seafar_init(ocean, 5, 240, INIT = "svd")
}
}
\author{
Tra Le and Katrijn Van Deun
}
