% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar_lasso.R
\name{seafar_lasso}
\alias{seafar_lasso}
\title{Title Sparse exploratory approximate factor analysis using LASSO penalty}
\usage{
seafar_lasso(
  data,
  nfactors,
  lambda,
  maxiter = 50,
  eps = 10^-4,
  initloadings,
  INIT = "mixed"
)
}
\arguments{
\item{data}{A data frame containing the dataset (standardized).}

\item{nfactors}{The number of factors.}

\item{lambda}{LASSO penalty tuning parameter.}

\item{maxiter}{Maximum number of iterations of the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{INIT}{Method to initialize loadings.}
}
\value{
Factor loading and factor score matrices.
}
\description{
Title Sparse exploratory approximate factor analysis using LASSO penalty
}
\examples{
\dontrun{
big5_seafa <- seafar_lasso(ocean, 5, lambda = 200, INIT = "svd")
}
}
