% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar_lasso.R
\name{seafar_lasso_multistart}
\alias{seafar_lasso_multistart}
\title{Multistart procedure for seafar with the LASSO penalty.}
\usage{
seafar_lasso_multistart(
  data,
  nfactors,
  lambda,
  maxiter = 50,
  eps = 10^-4,
  initloadings,
  INIT,
  nstarts
)
}
\arguments{
\item{data}{A data frame containing the dataset (standardized).}

\item{nfactors}{The number of factors.}

\item{lambda}{LASSO penalty tuning parameter.}

\item{maxiter}{Maximum number of iterations of the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{INIT}{Method to initialize loadings.}

\item{nstarts}{Number of starting values.}
}
\value{
\item{loadings}{The best estimated loading matrix.}
\item{scores}{The best estimated factor score matrix.}
\item{Lossvec}{A list of vectors of loss values of each starting value.}
\item{Loss}{A vector of loss values of the best starting value.}
\item{converged}{A scalar where 1 is converged and 0 is not converged.}
}
\description{
Multistart procedure for seafar with the LASSO penalty.
}
\examples{
\dontrun{
big5_result <- seafar_lasso_multistart(
  scale(big5),
  nfactors = 5,
  lambda = 10,
  INIT = 'random',
  nstarts = 50)
}
}
