% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar_multistart.R
\name{seafar_multistart}
\alias{seafar_multistart}
\title{Multistart procedure for seafar.}
\usage{
seafar_multistart(
  data,
  nfactors,
  C,
  maxiter = 50,
  eps = 10^-4,
  INIT,
  initloadings,
  orthogonal = FALSE,
  nstarts = 50,
  show_progress = FALSE
)
}
\arguments{
\item{data}{Data matrix of standardized items NxJ.}

\item{nfactors}{Number of factors Q.}

\item{C}{Number of nonzero loadings.}

\item{maxiter}{Maximum number of iterations of the AO procedure.}

\item{eps}{Convergence criterion based on difference in loss between iterates.}

\item{INIT}{Method to initializing loadings.}

\item{initloadings}{Initial loading matrix if available.}

\item{orthogonal}{Orthogonal or non-orthogonal factors, default is FALSE.}

\item{nstarts}{Number of starts.}

\item{show_progress}{Option of print progress bar, default is FALSE.}
}
\value{
\item{loadings}{The best estimated loading matrix.}
\item{scores}{The best estimated factor score matrix.}
\item{PVE}{A list of vectors of PVE of each starting value.}
\item{Loss}{A vector of loss values of the best starting value.}
}
\description{
Multistart procedure for seafar.
}
\examples{
\dontrun{
big5_result <- seafar_multistart(
  scale(big5),
  nfactors = 5,
  C = 240,
  INIT = 'random',
  orthogonal = TRUE,
  nstarts = 50)
}
}
