% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar_lasso.R
\name{summary.lasso_multistart}
\alias{summary.lasso_multistart}
\title{Display a summary of the results of \code{seafar_lasso_multistart()}.}
\usage{
\method{summary}{lasso_multistart}(object, disp = "loadings", nrow = 10, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'lasso_multistart'.}

\item{disp}{The default is \code{"loadings"} which returns only the estimated loadings matrix.
Otherwise if \code{"full"}, function returns both the estimated loadings and factor scores matrices.}

\item{nrow}{Number of rows of the loading matrix to be displayed, default is 10.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{seafar_lasso_multistart()}.
}
\examples{
\dontrun{
summary(big5_result, disp = "full")
}
}
