% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar_multistart.R
\name{summary.multistart}
\alias{summary.multistart}
\title{Display a summary of the results of \code{seafar_multistart()}.}
\usage{
\method{summary}{multistart}(object, disp = "loadings", ...)
}
\arguments{
\item{object}{Object of class inheriting from 'multistart'.}

\item{disp}{The default is \code{"loadings"} which returns only the estimated loadings matrix.
Otherwise if \code{"full"}, function returns both the estimated loadings and factor scores matrices.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{seafar_multistart()}.
}
\examples{
\dontrun{
summary(big5_result, disp = "full")
}
}
