% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seafar.R
\name{summary.seafar}
\alias{summary.seafar}
\title{Display a summary of the results of \code{seafar()}.}
\usage{
\method{summary}{seafar}(object, disp = "loadings", ...)
}
\arguments{
\item{object}{Object of class inheriting from 'seafar'.}

\item{disp}{The default is \code{"loadings"} which returns only the estimated loadings matrix.
Otherwise if \code{"full"}, function returns both the estimated loadings and factor scores matrices.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{seafar()}.
}
\examples{
\dontrun{
results <- seafar(X)
summary(results, display = "full")
}
}
